rename parser to parsers
because its spell is more correct and preferred way.
